+++ tailwind.config.ts
@@ -8,6 +8,14 @@ export default {
   ],
   theme: {
     extend: {
+      fontSize: {
+        'custom-15': [
+          '15px',
+          {
+            lineHeight: '26px'
+          }
+        ]
+      },
       colors: {
         background: "var(--background)",
         foreground: "var(--foreground)",
